mod: refactor handle events
